bug[delete joboffer]: - Delete Method not allowed at /joboffer/delete/id even through correct method was used

+++ frontend/src/pages/Stellenansicht.tsx
@@ -46,11 +46,21 @@ const Stellenansicht: React.FC = () => {
 
   // Funktion zum löschen der Joboffer
   async function deleteJoboffer(id: string | undefined) {
-    const response = await applicationTrackerApi.delete(
-      `http://localhost:8080/joboffer/delete/${id}`
-    );
-    if (response.data === 200) {
-      window.close();
+    if (id) {
+      console.debug(`Joboffer Löschen für Id: ${parseInt(id, 10)}`);
+      try {
+        const response = await applicationTrackerApi.delete(
+          `http://localhost:8080/joboffer/delete/${parseInt(id, 10)}`
+        );
+        console.debug(`Status Joboffer Löschen: ${response.status}`);
+        if (response.status === 200) {
+          window.close();
+        }
+      } catch (error) {
+        console.debug(
+          `Beim Löschen der Joboffer ist ein Fehler aufgetreten: ${error}`
+        );
+      }
     }
   }
